Overall:
  - NOTE: Some of this code is not shippable yet (particularly the document picker/store).
  - Builds with Xcode 4.6 with very few clang-sa warnings (remaining warnings might be clang bugs, but haven't narrowed them down to clear test cases yet)
  - Requires iOS 6 and Mac OS X 10.7
  - Removed iCloud support. Added OmniFileExchange framework, which is where the OmniPresence bits will be later (but not adding those yet).

Other highlights:

Configurations:
  - Sorted warning options in the same order as Xcode 4.6 lists them
  - Enabled a few more warnings
  - Require iOS 6
  - Removed some settings that Xcode 4.6 notes as deprecated.

Scripts:
  - BuildSVNVersionInclude avoids a rebuild when it can. Added support for svn 1.7.
  - BuildStringsFromTarget avoids a rebuild when it can.
  - CopyLibraryResources
    - Added an --exclude option for skipping resources that the running app doesn't need
    - Add support for DEVELOPER_USR_DIR containing spaces

OmniBase:
  - Fixed support for including backtraces in NSError userInfo dictionaries.
  - Added NSString-formatting attributes where appropriate.
  - Fixed some uses of objc_msgSend as function pointers, added utility functions for common cases (for example, OBCallVoidIMP).
  - More ARC support macros.
  - Added OB_HIDDEN for replacing __private_extern__ and OB_VISIBLE for cases where a symbol should be explicitly visible.
  - Added OBStrong{Retain,Release} for cases were you really need an object to stick around in ARC mode
  - Added OBInitializeDebugLogLevel helper for supporting getenv/NSUserDefaults style debug levels
  - Updated assertion macros to take a reason format and __VA_ARGS__.

OmniFoundation:
  - Moved some AppleScript helpers here from OmniOutliner/OmniFocus.
  - Various cleanup of public headers to support being imported in ARC-aware code.
  - Updated some files to be compiled with ARC, but most is still not.
  - Moved OFWeakRetain into OWF.
  - Added a OFWeakReference class for passing around or keeping lists of weak references.
  - OFController now uses ARC for observers via a new OFWeakReference class.
  - Fixed OFFastMutableData. 10.8 changed NSObject so that you can't keep a free list of objects via -dealloc.
  - Add OFOffsetMutableArray
  - Switched to build-time strings file generation via BuildStringsFromTarget.
  - Moved Keychain wrappers to this framework and updated the API some. Still more to do for certificate trust.
  - Added -[OFController shouldLogException:mask:] to allow OAController to filter out internal exceptions from Accessibility.
  - Added OFFilePresenterEdits, which helps deal with some of the oddities of NSFilePresenter notifications
  - Added OFHTTPHeaderDictionary, extracted from OWF
  - Added OFHTTPState and OFHTTPStateMachine, used to manage a graph of HTTP operations by transitions via HTTP status code
  - Added OFNetStateNotifier and OFNetStateRegistration, a pair of classes used to communicate availability of state changes across a group of members
    - The API here isn't final, but this the means by which OmniPresence clients notice edits on other devices and sync to pick them up
  - Removed OFImplementationHolder. Use blocks.
  - Added various new unit tests

OmniFileStore:
  - Extracted the guts of OFSDAVFileManager into OFSDAVConnection, which is fully async.
  - OFSDAVOperation now only uses the NSURLConnection operation queue scheduling and has no 'run' synchronously support
  - OFSFileManagerAsynchronousOperationTarget is gone, replaced by blocks that can be set on the operation
  - Updated OFSDAVFileManager's synchronous wrapper around OFSDAVConnection to not pump the runloop
  - Added basic support for LOCK/UNLOCK
  - Added support for ETag, lock token and missing predication in various cases
  - Added basic tests of WebDAV conformance, runnable at unit test time and server account validation time
  - Added a script to test up a local Apache test server
  - Made OFSDocumentStoreScope its own class instead of being an idea in OFSDocumentStore
  - Removed iCloud support.
  - Removed global file uniquification (in the style that iWork does).
  - NOTE: Some actions on the document store used to support the document picker aren't yet ported to the new API

OmniFileExchange:
  - New framework. This currently just maintains a registry of accounts.

OmniUI/iPad:
  - Split out all the document support into a new OmniUIDocument library
  - Added a first pass of accessibility support. Still plenty more to do, but this is a big step forward.
  - Lots of little UI/bug fixes to inspectors

OmniUI/iPad-Document
  - All the document based stuff lives here now, building in ARC mode
  - Reworked UI for maintaining cloud accounts. Still in progress.
  - Much better support for huge document sets w.r.t loading/managing previews
  - Support for document filters in the main document picker toolbar
  - NOTE: WebDAV-based import/export in the OmniUI document picker is currently broken

OmniAppKit:
  - OAInternetConfig asserts we have proper sandbox entitlements to send feedback mail
  - Added debugging helper for visualizing constraints on a view (see OAViewPicker and -visualizeConstraintsForPickedView:)
  - Fixed some issues with target selection vs. Powerbox.
  - Updated for OFWeakRetain going away
  - Updated toolbar item script support for sandboxing
  - Added OATrackingLoop, a block-based object to do mouse tracking.
  - Added Retina images.
  - Switched to build-time strings file generation via BuildStringsFromTarget.
  - Removed OAImageManager.

OmniInspector:
  - Switched to build-time strings file generation via BuildStringsFromTarget.
  - Updated to allow images to be pulled from the owning app
  - Added support for enabled bindings
  - Updated for Retina
  - Updated to allow an empty list of inspectors

OmniQuartz:
  - Improved vector rendering of layers

OmniSoftwareUpdate:
  - Switched to build-time strings file generation via BuildStringsFromTarget.
  - Retina support
  - Updated support for a button to take you to a web page

OWF:
  - Moved OWFWeakRetain here from OmniFoundation (which will prefer ARC __weak in the future)
  - Moved much of OWHeaderDictionary to OmniFoundation as OFHTTPHeaderDictionary
  - Ported more code to 64-bit
  - Rewrote a bunch of OWSGML* to use blocks instead of OFImplementationHolder, which has been removed

OIF:
  - Updated for new location of OWFWeakRetain stuff

+++ Frameworks/OWF/Content.subproj/OWCompoundObjectStream.h
@@ -1,4 +1,4 @@
-// Copyright 1999-2005, 2010 Omni Development, Inc.  All rights reserved.
+// Copyright 1999-2005, 2010, 2013 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -16,14 +16,14 @@
     OWAbstractObjectStream *framingStream;
     OWAbstractObjectStream *interjectedStream;
 
-    unsigned int interjectedAtIndex;
+    NSUInteger interjectedAtIndex;
 }
 
 /* a convenience method */
 + (OWObjectStreamCursor *)cursorAtCursor:(OWObjectStreamCursor *)aCursor beforeStream:(OWAbstractObjectStream *)interjectMe;
 
 /* designated initializer */
-- initWithStream:(OWAbstractObjectStream *)aStream interjectingStream:(OWAbstractObjectStream *)anotherStream atIndex:(unsigned int)index;
+- initWithStream:(OWAbstractObjectStream *)aStream interjectingStream:(OWAbstractObjectStream *)anotherStream atIndex:(NSUInteger)index;
 
 /* raises an exception if aStream is not a (possibly indirect) member of this compound object stream */
 - (NSUInteger)translateIndex:(NSUInteger)index fromStream:(OWAbstractObjectStream *)aStream;
