+++ Frameworks/OWF/Content.subproj/OWDataStreamFilterCursor.h
@@ -1,4 +1,4 @@
-// Copyright 2003-2005, 2010 Omni Development, Inc.  All rights reserved.
+// Copyright 2003-2005, 2010, 2013 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -22,7 +22,7 @@
 
 // Subclass' responsibility.
 - (void)processBegin;
-- (void)fillBuffer:(void *)buffer length:(unsigned)bufferLength filledToIndex:(unsigned *)bufferFullp;
+- (void)fillBuffer:(void *)buffer length:(NSUInteger)bufferLength filledToIndex:(NSUInteger *)bufferFullp;
 
 // A concrete subclass of OWDataStreamFilterCursor must provide implementations for the following methods:
 //    -fillBuffer:length:filledToIndex:
