+++ Frameworks/OWF/Pipelines.subproj/OWContentInfo.h
@@ -1,4 +1,4 @@
-// Copyright 1997-2005, 2010 Omni Development, Inc.  All rights reserved.
+// Copyright 1997-2005, 2010, 2013 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -40,7 +40,7 @@
         unsigned int isHeader:1;
     } flags;
     NSLock *flagsLock;
-    int workToBeDoneIncludingChildren;
+    size_t workToBeDoneIncludingChildren;
 
     OFMessageQueueSchedulingInfo schedulingInfo;
 }
@@ -84,8 +84,8 @@
 - (NSArray *)childTasks;
 - (OWTask *)childTaskAtIndex:(NSUInteger)childIndex;
 - (NSUInteger)childTasksCount;
-- (int)workDoneByChildTasks;
-- (int)workToBeDoneByChildTasks;
+- (size_t)workDoneByChildTasks;
+- (size_t)workToBeDoneByChildTasks;
 - (void)calculateDeadPipelines:(NSUInteger *)deadPipelines totalPipelines:(NSUInteger *)totalPipelines;
 
 - (void)addChildFossil:(id <NSObject>)childFossil;
